requirements.txt
# Core
scrapy==2.12.0
pymongo==4.6.1
greenstalk==2.0.2
python-dotenv==1.0.0
psutil==5.9.5

# JavaScript rendering
splash==3.5
playwright==1.40.0
scrapy-playwright==0.0.33

# Proxy handling
scrapy-rotating-proxies==0.6.2

# HTTP utilities
requests==2.31.0
urllib3==2.0.6

# Data processing
lxml==4.9.3
w3lib==2.1.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
retrying==1.3.4
tabulate==0.9.0

# Testing
pytest==7.4.0
pytest-mock==3.11.1
pytest-cov==4.1.0

# Linting/formatting
flake8==6.1.0
black==23.7.0
isort==5.12.0